Custom field naming handling and atoms now converted to/from binaries
It is now possible to specify remote function in 'Options' of 'gen_encode/4' and 'gen_decode/4' via 'name_handler' option and this function will be used to convert record's field names to binary passed to 'mochijson2'.
Also all atom values are now converted to binary before passing to 'mochijson2' (and vise versa).