Remove support for `mochijson2` from dev for now
All related files are moved to `mochijson2` branch.
If there is a way to do conditioned `rebar` i.e. use a flag to specify that support for `mochijson2` is needed (and dependency is required - that is why it has been moved for now) which is turned off by default, those files will be moved back